refactor(site): 统一插件目录结构
