+++ site/plugins/icons-manifest/index.js
