+++ site/plugins/tdoc/md-to-vue.js
@@ -11,20 +11,20 @@ export default function mdToVue(options) {
     `;
   }
 
-  const sfc = `
-    <template>
-      <td-doc-content ref="tdDocContent" page-status="hidden">
-        ${
-          mdSegment.tdDocHeader
-            ? `
+  const docHeader = mdSegment.tdDocHeader
+    ? `
           <td-doc-header
             slot="doc-header"
             ref="tdDocHeader"
             spline="${mdSegment.spline}"
           >
           </td-doc-header>`
-            : ''
-        }
+    : '';
+
+  const sfc = `
+    <template>
+      <td-doc-content ref="tdDocContent" page-status="hidden">
+        ${docHeader}
         <div name="DOC">${mdSegment.docMd}</div>
         <td-doc-footer slot="doc-footer"></td-doc-footer>
       </td-doc-content>
@@ -86,7 +86,7 @@ function customRender({ source, md }) {
 
   const mdSegment = {
     ...pageData,
-    docMd: md.render.call(md, `${pageData.toc ? '[toc]\n' : ''}${content}`).html,
+    docMd: md.render(`${pageData.toc ? '[toc]\n' : ''}${content}`).html,
   };
 
   return mdSegment;
